Refactor : 레주메에도 인증/인가 로직 추가

+++ src/services/resumes.service.js
@@ -13,10 +13,7 @@ export class ResumesService {
 		const resumes = await this.resumesRepository.findAllResumes();
 
 		if (!resumes) {
-			throw new ApiError(
-				404,
-				`이력서 데이터가 없습니다.`,
-			);
+			throw new ApiError(404, `이력서 데이터가 없습니다.`);
 		}
 
 		resumes.sort((a, b) => {
@@ -71,10 +68,7 @@ export class ResumesService {
 		const resume = await this.resumesRepository.findResumeById(resumeId);
 
 		if (!resume) {
-			throw new ApiError(
-				404,
-				`아이디가 ${resumeId}인 이력서가 존재하지 않습니다.`,
-			);
+			throw new ApiError(404, `해당 이력서가 없습니다.`);
 		}
 
 		return {
@@ -87,43 +81,49 @@ export class ResumesService {
 		};
 	};
 
-	updateResume = async (resumeId, title, contents, statusCode) => {
+	updateResume = async (
+		resumeId,
+		userId,
+		title,
+		contents,
+		statusCode,
+		authCode,
+	) => {
 		const resume = await this.resumesRepository.findResumeById(resumeId);
 
-		if (!resume) throw new ApiError('존재하지 않는 이력서입니다.');
+		if (!resume) {
+			throw new ApiError(404, `해당 이력서가 없습니다.`);
+		}
 
-		await this.resumesRepository.updateResume(resumeId, userId, title, content);
+		if (resume.userId != userId && authCode !== 'admin') {
+			throw new ApiError(403, `본인의 이력서만 수정 가능합니다.`);
+		}
 
-		const updatedResume = await this.resumesRepository.findResumeById(resumeId);
+		await this.resumesRepository.updateResume(
+			resumeId,
+			title,
+			contents,
+			statusCode,
+		);
 
 		return {
-			resumeId: resume.resumeId,
-			userId: resume.userId,
-			title: resume.title,
-			contents: resume.contents,
-			createdAt: resume.createdAt,
-			updatedAt: resume.updatedAt,
+			message: '정상 수정되었습니다.',
 		};
 	};
 
-	deleteResume = async (resumeId) => {
+	deleteResume = async (resumeId, userId, authCode) => {
 		const resume = await this.resumesRepository.findResumeById(resumeId);
 		if (!resume) {
-			throw new ApiError(
-				404,
-				`존재하지 않는 이력서입니다.`,
-			);
+			throw new ApiError(404, `존재하지 않는 이력서입니다.`);
+		}
+		if (resume.userId != userId && authCode !== 'admin') {
+			throw new ApiError(403, `본인의 이력서만 삭제 가능합니다.`);
 		}
 
 		await this.resumesRepository.deleteResume(resumeId);
 
 		return {
-			resumeId: resume.resumeId,
-			userId: resume.userId,
-			title: resume.title,
-			contents: resume.contents,
-			createdAt: resume.createdAt,
-			updatedAt: resume.updatedAt,
+			message: '정상 삭제되었습니다.',
 		};
 	};
 }
